+++ src/CodePunk.Console/Commands/Modules/ICommandModule.cs
@@ -1,0 +1,8 @@
+using System.CommandLine;
+
+namespace CodePunk.Console.Commands.Modules;
+
+internal interface ICommandModule
+{
+    void Register(RootCommand root, IServiceProvider services);
+}
